Adds db initializer, reverts location coordinates to lat, log from Geometry

+++ app/models/location.model.js
@@ -5,8 +5,8 @@ module.exports = (sequelize, Sequelize) => {
     const Location = sequelize.define("location", {
         id: { type: Sequelize.INTEGER, primaryKey: true, autoIncrement: true },
         name: { type: Sequelize.STRING, allowNull: false },
-        latitude: { type: Sequelize.INTEGER, allowNull: true, defaultValue: null, validate: { min: -90, max: 90 } },
-        longitude: { type: Sequelize.INTEGER, allowNull: true, defaultValue: null, validate: { min: -180, max: 180 } },
+        latitude: { type: Sequelize.DECIMAL, allowNull: true, defaultValue: null, validate: { min: -90, max: 90 } },
+        longitude: { type: Sequelize.DECIMAL, allowNull: true, defaultValue: null, validate: { min: -180, max: 180 } },
         address_complete: { type: Sequelize.STRING, allowNull: true, defaultValue: null },
         street_1: { type: Sequelize.STRING, allowNull: true, defaultValue: null },
         street_2: { type: Sequelize.STRING, allowNull: true, defaultValue: null },
